Extract object anim import to own file
